+++ qiskit/pulse/__init__.py
@@ -132,12 +132,13 @@ Exceptions
 
 from .channels import (DriveChannel, MeasureChannel, AcquireChannel,
                        ControlChannel, RegisterSlot, MemorySlot)
-from .commands import AcquireInstruction, FrameChange, PersistentValue
+from .commands import FrameChange, PersistentValue
 from .configuration import LoConfig, LoRange, Kernel, Discriminator
 from .exceptions import PulseError
 from .instruction_schedule_map import InstructionScheduleMap
 from .instructions import Acquire, Instruction, Delay, Play, ShiftPhase, Snapshot, SetFrequency
 from .interfaces import ScheduleComponent
-from .pulse_lib import SamplePulse, Gaussian, GaussianSquare, Drag, ConstantPulse, ParametricPulse
+from .pulse_lib import Gaussian, GaussianSquare, Drag, ConstantPulse
+from .types import SamplePulse, ParametricPulse, Pulse
 from .pulse_lib.samplers.decorators import functional_pulse
 from .schedule import Schedule
