+++ test/python/pulse/test_samplers.py
@@ -18,7 +18,7 @@
 import numpy as np
 
 from qiskit.test import QiskitTestCase
-import qiskit.pulse.commands as commands
+from qiskit.pulse.types import SamplePulse
 import qiskit.pulse.pulse_lib.samplers as samplers
 
 
@@ -46,7 +46,7 @@ class TestSampler(QiskitTestCase):
         reference = np.array([0.1, 0.2], dtype=np.complex)
 
         pulse = left_linear_pulse_fun(duration, m=m, b=b)
-        self.assertIsInstance(pulse, commands.SamplePulse)
+        self.assertIsInstance(pulse, SamplePulse)
         np.testing.assert_array_almost_equal(pulse.samples, reference)
 
     def test_right_sampler(self):
@@ -58,7 +58,7 @@ class TestSampler(QiskitTestCase):
         reference = np.array([0.2, 0.3], dtype=np.complex)
 
         pulse = right_linear_pulse_fun(duration, m=m, b=b)
-        self.assertIsInstance(pulse, commands.SamplePulse)
+        self.assertIsInstance(pulse, SamplePulse)
         np.testing.assert_array_almost_equal(pulse.samples, reference)
 
     def test_midpoint_sampler(self):
@@ -70,7 +70,7 @@ class TestSampler(QiskitTestCase):
         reference = np.array([0.15, 0.25], dtype=np.complex)
 
         pulse = midpoint_linear_pulse_fun(duration, m=m, b=b)
-        self.assertIsInstance(pulse, commands.SamplePulse)
+        self.assertIsInstance(pulse, SamplePulse)
         np.testing.assert_array_almost_equal(pulse.samples, reference)
 
     def test_sampler_name(self):
@@ -81,7 +81,7 @@ class TestSampler(QiskitTestCase):
         left_linear_pulse_fun = samplers.left(linear)
 
         pulse = left_linear_pulse_fun(duration, m=m, b=b, name='test')
-        self.assertIsInstance(pulse, commands.SamplePulse)
+        self.assertIsInstance(pulse, SamplePulse)
         self.assertEqual(pulse.name, 'test')
 
     def test_default_arg_sampler(self):
@@ -92,5 +92,5 @@ class TestSampler(QiskitTestCase):
         reference = np.array([0.1, 0.2], dtype=np.complex)
 
         pulse = left_linear_pulse_fun(duration, m=m)
-        self.assertIsInstance(pulse, commands.SamplePulse)
+        self.assertIsInstance(pulse, SamplePulse)
         np.testing.assert_array_almost_equal(pulse.samples, reference)
